+++ src/components/DetailsCard/DetailsCard.scss
@@ -11,14 +11,13 @@
   box-shadow: inset 0px -2px 0px rgba(0, 0, 0, 0.25);
   border-radius: 0px 0px 16px 16px;
   width: 100%;
-  height: 140px;
+  min-height: 140px;
 }
 
 .details-card {
   background: #ffffff;
   border-bottom: 2px solid #dfe1e3;
   box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.04), 0px 4px 8px rgba(0, 0, 0, 0.06);
-  min-height: 110px;
   border-radius: 1rem;
   margin: 0 20px;
   padding: 0.75rem;
@@ -48,9 +47,9 @@
     text-align: center;
   }
   .details__subtitle {
-    font-family: ui-rounded, 'Rubik', system-ui, sans-serif;
-    font-weight: 400;
-    font-size: 0.875rem;
+    font-family: system-ui, ui-rounded, 'Rubik', sans-serif;
+    font-weight: 500;
+    font-size: 1rem;
     letter-spacing: 0.04em;
     color: v.$washed-gray;
   }
