+++ src/pages/Profile/Profile.js
@@ -1,4 +1,3 @@
-
 import './Profile.scss';
 import { useParams } from 'react-router-dom';
 import { useEffect, useState } from 'react';
